Made coordinates double precision.
Not sure if this is necessary. Related to Cork model issues, which had completely different cause, haha!

+++ source/src/sfincs_crosssections.f90
@@ -20,8 +20,8 @@ contains
    real*4    :: dphi
    character(len=256) :: cdummy
    !
-   real*4, dimension(:),   allocatable :: xcrs
-   real*4, dimension(:),   allocatable :: ycrs
+   real*8, dimension(:),   allocatable  :: xcrs
+   real*8, dimension(:),   allocatable  :: ycrs
    real*4,  dimension(2)                :: xp
    real*4,  dimension(2)                :: yp
    integer, dimension(:),   allocatable :: uv_indices
